refactor: move GGUFTokenizer to llm-core, add TokenizerFactory
Phase 3 of the unified pipeline plan. Tokenization is now a standalone
pipeline stage in llm-core, independent of any specific runner.

- Move GGUFTokenizer from kllama to llm-core/tokenizer package
- Add typealias in kllama for backwards compatibility
- Create TokenizerFactory with fromGGUF(), fromTokenizerJson(), fromHuggingFace()
- Add skainet-io-gguf and kotlinx-io-core dependencies to llm-core
